Prevent workflows from running in parallel on the same event
This patch fixes a bug which caused the workflow service to only check
the first 20 workflows with a specific media package identifier for
running workflows before starting a new one. This means if 20 had
already run in the past, you could happily start workflows on the same
media package in parallel and no one would prevent that.