Fix isActive check for workflows
This patch fixes the `isActive` check for workflows from the workflow
utils which should be the inverse of `isTerminated` but was actually
missing the `failing` state.

This also marks this method as deprecated since with the existence of
`status.isTerminated()`, there is no real benefit in having a separate
utils method for this check.