Only allow access with write permissions
This patch makes the editor backend check if a user accessing the
standalone video editor has write permissions on the event to be
edited. This avoids all kind of confusion and several errors resulting
in errors when users try to save things they have edited but are not
actually allowed to change. Better to be upfront about the missing
permissions.

This means that users cannot misuse the editor as preview player any
longer but… we have better players for that.