Provide a minimal "No Op" workflow for debugging and maintenance
This can be useful sometimes, for example when parts of our indices
concerning workflows get out of sync with the database, and you see
inconsistent information about the state of a workflow in the events
table vs. the event details. In a case like that, you can just run
some workflow using the `/workflow/start` API to force Opencast
to go through all the proper channels to update the database
and the indices in a (hopefully) consistent manner.

Having a workflow for that purpose handy is thus often useful,
so admins don't have to have it floating around somewhere external,
or come up with the definition from scratch every time. You can just
specify `noop` as a parameter to the `/workflow/start` endpoint.
Also, this workflow intentionally has no `tags` so that it doesn't
show up anywhere in the UI to confuse people.