Don't update series name of events in index twice
Right now, the update of events in the index is done twice when
triggered by a series metadata change: Once directly in the series
service when updating the series in the index, and once asynchronously
by the AssetManagerUpdatedEventHandler when doing a snapshot. The first
approach is faster, but will only update the index, not the actual event
data. So if something goes wrong during the snapshot, this will lead to
inconsistency. It's also a bit of a code smell in my opinion to do this
twice. So I'd like to remove this, even though it might result in slower
performance, because the index should represent the actual state of the
data, not the state it will hopefully soon be in. But I'm open to
criticism.