Added comment fetching to event fetching
The "/event.json" admin ui endpoint now adds comments to the event. This replaces comment fetching in the frontend, which caused serious load and made the notes column unusable on larger systems.

Supposedly a good solution to this problem would be to add comments to the index, but that is beyond my capabilities atm.