Fix LTI Tool Build
In TypeScript 4.4.3 the type declarations for `top` has changed to
`WindowProxy | null`. That is why pull request #2956 broke the build
once it was merged into `develop`. `develop` already uses the new
TypeScript version. The result is:

```
Failed to compile.

/home/lars/dev/opencast/modules/lti/src/components/Series.tsx
TypeScript error in /home/lars/dev/opencast/modules/lti/src/components/Series.tsx(184,9):
Object is possibly 'null'.  TS2531

    182 |
    183 |     annotateEpisodeCallback(id: string) {
  > 184 |         window.top.location.href = "/annotation-tool/index.html?id=" + id
        |         ^
    185 |     }
    186 |
    187 |     downloadEventCallback(track: Track) {

```

According to the specification, `window.top` may never be `null` in a
bowser context. That is why it should be safe to just tell TypeScript to
ignore that this type could be `null`.

This fixes the build.