Add `getAccessControlList` to `SearchResultItem`
The ACL information is available in the search index in non-XML form.
Code using the search service might be interested in that ACL without
needing to load an attachment and parse XML. So it makes sense to
provide this information as part of `SearchResultItem`.

Unfortunately, to make the `SearchService` remote impl work (which uses
the search API), the access control list has to be serialized and
exposed in that API. The ACL in the index also always contains the
organization admin role for read and write. This is also now exposed in
the API.