Clarify development dependency container documentation
It's true that no persistence is configured in the compose files, but
the container images used declare `VOLUME`-s for  their data
directories. Without further configuration in the compose file, this
will lead to the data being stored in anonymous volumes. To actually
get a fresh start, you have to recreate the containers, not just
restart them