Minor Workflow Changes
This patch contains a few minor modifications to the new workflows.
The changes are mainly:

- Documentation/Upgrade guide
- Workflow descriptions
- Removed or collapsed some tag operations
- Removed some unnecessary operations from fast

There is still room for optimizations. For example, you could remove
some more tag operations and there is an exceeding amount of analyze
tracks, but that's maybe something for later.

Additionally, we may want to remove the timelinepreview and cover-image
operations. But that needs additional discussions.