Exclude administrators from artificial limit in `SolrRequester`
This artificial limit of 2000 was added in PR #1092, this commit:
608720f4add3706e7bb2c3f9b67a1589420bca0a

It was added because requesting the search API without limit would try
to retrieve all events, which could exhaust the server's memory.

Given that this limit was completely undocumented, this is a surprising
behavior for the `SearchService`. In my opinion it would be best to
only have such an artificial limit for external APIs that normal users
can use. But unfortunately, the search API is also used internally by
the `SearchServiceRemoteImpl`. So without massively rewriting the
remote impl, I don't see how one could really improve upon it.

This commit only lifts the limit for admins. This still protects against
DOS attacks by non-admin users while still making the SearchService
more useful to some other Opencast modules.