Let Encode Handle HLS
The internal code of the encode operation will now pick up HLS playlist
but this is unfortunately not handles later in the code, causing the
inspect operations to fail. This fixes the issue by looking out for HLS
playlists, ensuring the file references in there are correct during
inspection.

This also changes the way inspections are run, switching from a
sequential run on all files to running the inspection in parallel, like
we do with the other operations.

This shouldn't change the default behavior of the encode operations in
any way.