Fixing security issue where system wide digest credentials were being inappropriately sent to unauthenticated servers.
It looks like someone tried to fix this already with the org.opencastproject.download.source config key, however the
logic was incorrect and the credentials were being sent regardless.  This patch fixes the logic and adds tests to try
to ensure that the behaviour can't break again.

Filed as GHSA-hcxx-mp6g-6gr9