Extend Metadata Extraction in Inbox
Opencast's inbox allows you to add zipped media packages which may
contain arbitrary metadata but are very complex or to add simple files,
which file names will end up as the new event's title. It would be nice
if it would be possible to extract a few more basic metadata from the
filename.

File names like these are commonly generated by capture devices:

```
2020-02-07-hs1-Introduction to Biology.mp4
 |    |  |  |  |                     |
year  |  |  |  +----------+----------+
   month |  |             |
        day |           title
        location
```

To improve the situation, and to be as flexible as necessary, we would
like to allow for specifying regular expressions as matchers for a few
basic metadata.

This resolves #3239