Inbox may fail at startup
The inbox service may start up prior to the organization configuration
being loaded. Since the inbox sets an internal security context to run
as a specific user in a specific organization, this can lead to the
inbox not finding that user or organization and thus failing to
configure an inbox.

This patch makes the inbox wait for the organization and user to show up
instead of failing. This means that startup order does not matter any
longer.

This fixes #3243
This is based on #3248