Fix posible type change in external API (#3231)
PR #3148 introduced a possible type change in the external API from int
to string for the duration field of tracks in case the field is `null`.
As discussed, this PR changes the return value to JSON `null` instead.

Note that the behavior is still inconsistend when it comes to durations:

* `{eventId}/media` will omit the field
* Event duration has a type change to `""`
* this changes Track duration to `null`

This patch leave the behavior of `{eventId}/media` as this is done with
all the fields. The Event duration though now also returns JSON `null`
as default value.