Fix Scheduler Conflict Checks
This patch fixes the scheduler's conflict checks which previously broke
spectacularly if someone already ingested media to one of the found
scheduled events.

The scheduler woould then find the conflict but would fail to retrieve
the media package since it wouldn't be the owner of that event in the
asset manager any longer. That does not make a lot of sense for a
conflict check where you just want to get the information.