Fix REST Test Netwprk Bindings
Using the REST tests environment, Opencast's tests would randomly fail
with ports seemingly being already in use, causing the bind of the test
server to fail.

This patch changes the test environment behavior to select a random port
and the directly trying a bind instead of selecting a port in advance.
This should cause the test server to select another port if the bind
fails.

This (hopefully) fixes #2098