Use ffprobe to get additional metadata
In addition to the metadata extraction from the file name, this patch
allows users to have ffprobe run on the media files to get some
additional metadata from the media container.

The extracted metadata are creation time and duration, which may help to
match scheduled events.

With the media file now having a duration, there is now also an
additional option allowing to filter out some events which are barely
conflicting (imaging two back to back recordings and the second one
starts a minute early).