Handle missing common ui adapters for tenants
Currently tenants other than the default tenant don't actually have
common metadata configured (or rather, it's not labeled as common).
Since this causes NPEs with the most recent changes to the index
service, we fall back to the adapter for the default tenant now,
similarly to how it worked before. This way, no config changes are
necessary for this fix to work.