Revert order of common and extended metadata in admin ui event endpoint
This order was changed a while ago, but it's actually important for the
common adapter to overwrite any extended metadata adapters with the same
flavor in case of multi tenancy, where it's currently recommended to have
a copy of the common metadata config, but not label it common. Since the
common adapter gets some special treatment by the backend like renoving
some fields prior to being sent to the frontend, we should use it over the
"not-common" tenant-specific ones until there is a more general fix for
the whole situation. This means that the copies for the other tenants
are actually probably unnecessay right now, at least for the admin ui.