Fix Javadocs
With Opencast 11 our javadocs broke. In part this was due to the migration from JDK 8 to 11,
but the majority of those issues have already been corrected by this point. The current
issue, which this commit resolves, is that various Lucene packages (from modules/solr) are
not found by javadocs when the aggregation step happens. We have to manually add these to
the javadoc plugin to make sure that it can find them.

README: When testing aggregation bugs you can massively speed up testing by using the
javadocs.sh file generated in target/site/apidocs (from project root). This file disappears
as part of the javadoc generation process, so you won't find it unless something goes wrong :)

Fixes #2085