Capture agent calendar as JSON
This patch provides the capture agent calendar in JSON format. This
allows agents to talk to Opencast and get all necessary responses as
JSON which makes handling these data much simpler.

The new endpoint `/recordings/calendar.json` returns data like this:

```json
[
  {
    "eventId": "95b7b721-5d42-4073-9222-634e5cfd8aaa",
    "agentId": "a",
    "startDate": "Jan 19, 2022, 11:30:00 PM",
    "endDate": "Jan 19, 2022, 11:55:00 PM",
    "presenters": [],
    "workflowProperties": {
      "straightToPublishing": "true"
    },
    "agentConfig": {
      "capture.device.names": "",
      "org.opencastproject.workflow.definition": "fast",
      "org.opencastproject.workflow.config.straightToPublishing": "true",
      "event.location": "a",
      "event.title": "test"
    },
    "recording": {}
  }
]
```

Which is a serialized list of the scheduler's TechnicalMetadata class.