Remove Tracked Host Bundles
This patch adds a new REST endpoint to remove all tracked bundles for a
given host.

Normally, this happens automatically when a node is shut down or
restarted to have a global overview of what is running in a cluster. In
some cases, e.g. a machine is discarded and Opencast was not properly
shut down, a host is left in the table though.

Having an old host in that table only causes Opencast to report
inconsistent versions, nothing else. Still, before this patch it was hard
to get rid of that warning and could lead to confusion.

The new endpoint is accessible to admins only. They hopefully know what
they are doing when removing a host. But even if they remove an existing
host, the only side-effect is the version not being displayed properly
and that would auto-repair once Opencast has been restarted.