Finish renaming file using the title
This patch finishes the rename-files workflow operation handler with the
limitation that files are now always automatically renamed using the
event title.

The workspace is not used to access the files, assuring that Opencast
can get access, no matter where the data is stored. If it can get
access, of course.