Fix frontend-no-prebuilt profile
This patch fixes the frontend-no-prebuilt maven profile which can be
used to have maven run the local npm version instead of downloading one.
That has the big advantage of not having to remove any `none*` folders
when you switch versions.

To try this, run:

```
❯ mvn clean install -Pdev,frontend-no-prebuilt
```

The profile did exist for a long time but was broken due to some tasks
being executed in the wrong order like trying to copy the frontend to
the target folder before actually building it.