Make duration in Tobira API non-null
Having this field nullable ripples through the whole Tobira codebase
and adds lots of special cases everywhere. We decided to ignore this
special case and just use a dummy duration of 0 instead. I know, I know,
this is bad of course, in some sense. But we can simply consider it a
bug if an event does not have a duration. Opencast should know and can
know. In our large test data set, none of the events have an unknown
duration. And what's the alternative? Having it nullable everywhere,
in Tobira's frontend, backend and DB. This adds a lot of code for
nothing basically. And if the type system doesn't catch it, it can
easily introduce weird bugs. So we decided to accept this slight hack
in order to simplify other parts. Worst case: Tobira shows a 0 duration
for some events.