Fix workflow definition installation
When installing workflow definitions there is a chance that a workflow
definition which fails to load will result in stalling the boot of
further services. This happens when the faulty workflow definition is
loaded last and the `ReadinessIndicator` service is not registered.
This rearranges the code to make sure the service is always registered
when the last workflow definition was handled.