Fix bug with adding a series when a deleted one still lingers
`this.deleteSeries` only marks the series as deleted again, not actually
removing it. This will clash if a series with the same ID is added again
as it will cause a unique constraint violation. Instead, we completely
remove it here. Unfortunately, without the `flush` it didn't work: the
unique violation was still triggered. Whether `flush` is the best
solution: 🤷, I hardly know stuff about JPA.

This was noticed by Katrin:
https://github.com/opencast/opencast/pull/3076#discussion_r839548962