Adds a new endpoint to the External API Events endpoints. It allows uploading a track to an event by sending the updated media package to the archive. It does not start a workflow.
The motivation for this feature is to offer an external endpoint that allows uploading subtitles as tracks to an event after event creation and without access to the Admin UI. However, the endpoint is currently not restricted to subtitle files or the "captions/*" flavor, mainly because it was easier to code while also giving the endpoint additional functionality beyond its originally intended use case.

The endpoint does currently not place a restriction on which file types can be uploaded. Maybe it should?