Out of video image extraction should not fail
Some operations in Opencast make use of image extraction on given times in video, e.g. text analysis, seekbar preview. In some cases the video track is shorter then audio track (a few milliseconds to some seconds). In this case the image extraction will fail and the whole operation also fail. This patch will fix this by overlaying the video frame on a black background. If you try to extract a frame outside of the video track, you will get a black (background) image. The (positive) downside is the fixed size of resulting images.