Configure live schedule module for test deployment
This allows live events to be published (and thus be read by Tobira).
Currently we don't have our own streaming setup, so we use external
streams. Due to the way the streaming URL is configured for the module,
it's easiest to just map those external streams to localhost via nginx.