Add missing tags in encode WOH
Tags for different output files were missing in encode WOH, because
the comparison of the output file name to a specified suffix stopped
working after 19c0940c.
The reason for this is that the files were renamed not including
the specified suffix, and comparison for the suffix as a condition
for adding a tag is used.
Fixes the issue by remembering the original file name from encoding.

Resolves #3639