This is a horrible hack, but...
We need to be able to ensure that all of the handlers are up and connected prior to system startup, because otherwise
we can end up losing full or partial messages.  The handlers are dynamic bindings, and have to be to prevent startup
deadlocks - this is sufficiently complex that a large refactor of the search service is required to resolve it.  In
the meantime, this horrible hack enforces handler presence by hardcoding(!) the correct count and requiring that we
have that many handlers before proceeding.  This has many, obvious pitfalls, but since search will hopefully be
refactored soon I think this is the best course of action.  At least for now.