Fixed Bug in Metadata Tabs - additional fix
Previously in Metadata Tabs for fields which have drop-downs for value selection, while they are editable, the key of the value was displayed for the selected value instead of the name of the value, when in non-editable mode. This has been fixed, now the translated name of the selected value is displayed, even, when the field is non-editable.