No Elasticsearch Use In Ingest
The ingest service searches for series titles starting with pull
request #3586.  This causes trouble when combined with the removal of
the Solr index which now makes this code fall back to the Elasticsearch
index instead.

Unfortunately, the Elasticsearch modules are not included in the ingest
distribution (#3797) which will cause the distribution to fail.

Simply including it works in theory, but causes trouble since our
infrastructure documentation states that Elasticsearch may be available
on the admin only and can thus, for example, be run with no encryption
or authentication to make deployments easier. All this would need to
change, making this quite messy.

This patch fixes the issue by using a generated series identifier for
which we need to full text search instead.