Initial pass at fixing the tests.  Logic below:
The tests were deadlocking on the .wait() calls, which were quite old.  I suspect that the previous workflow logic did enough stuff between a state change, and the wait() in the test that the wait happened *before* the state change (the expected behaviour).  With the newer implementation logic, the state change happens *before* the wait, and thus the wait deadlocks waiting on an event which has already happened.