Indexing only latest workflows for all events (#3775)
Added 2 named queries and the corresponding methods within the workflowDatabaseService to count and read the latest/active workflows of each event. Replaced the list of the repopulate function within WorkflowServiceImpl containing all workflows with a new one containing only the active/latest workflows. The rest of the indexing procedure remains the same.