Improved workflow data processing
Workflow data is now processed when doing a rebuild without creating a new map and overwriting the old workflows within it until the latest workflows are fetched. Instead of the data is sorted by mp ID and wf ID and only the first (-> latest) instance of each mp ID will be updated. Also removed the step for changing the security context because we don't need it any longer within the new procedure.