Use Local Index References
Services now track their index on their own and use it to push updates
into the index if things change. For example, when a new workflow is
started.

For reindexing, an external index was still passed to the service
though. That's pretty weird since that could mean that the same service
would write to completely different indexes in different situations,
but expecting that the data would end up at the same place.

In practice this did not really matter since the same index was passed
to the `repopulate()` method which was tracked internally anyway. Still,
it makes more sense to always use the tracked index.