Updates karma-jasmine to 5.1.0
Closes #3877

Removed an empty statement in loginControllerSpec.js.

Tried to fix up PlayerAdapterFactory tests, but running the same spec for multiple services is apparently something that no other jasmine users do, so this is kinda iffy. Although previously these tests did not run at all and now they at least run for the default factory, so that's still a net plus.