Generate Safe Names in Working File Repository
This patch moves the code for generating safe file names from the static
PathSupport class to the working file repository in preparation for
making this configurable. Developers can now access this functionality
either via the working file repository or via the workspace, although
most of the working file repository and workspace methods ensure safe
file names anyway and there is often no need to call this separately.

Most of the changes in this patch stem from changes to unit tests which
mock the workspace and now needs to be mocked differently and from
removing the unused remote implementation of the working file
repository, so that I do not have to modify it.