Add navigation shortcuts in admin ui modals
Key combinations:
 'alt + enter' for 'next' or
 - if applicable - the 'submit' button,
and:
 'alt + backspace' for the 'previous' button.

They are implemented in wizardDirective.js and call the directive's 'toTab' method.
'toTab' expects a specific event which is tied to and emitted by
pressing the 'next' or 'previous' button.  As it is not possible to pass
the expected event when using the key-combinations, the 'toTab' method
was minimally extended to feature an alternative and optional parameter
called 'altTargetStepName'.
This expects a string which can either be 'next' or 'previous', and is
passed by the key-combinations. Every other 'toTab' call should work as
it did before without any changes.

US translations for the hotkey cheatSheet are included.