Skip broken items instead of returning 500
Before this, any exception in `Item` constructors (where we retrieve the
information about events/series) would lead to us returning 500 from
the Tobira API. Over time we noticed, that these exception sadly happen
from time to time because events or series are just broken. And don't
have data we assume they have.

While this runs the risk of administrators never noticing that an event
is broken and thus not synchronized with Tobira, we just log an error
and skip the item. We don't want to outright make synchronization
impossible when there is a broken item. This should be combined with
proper monitoring to notice whens something like this goes wrong.