Fail on configuration error
This patch lets the video editor fail on configuration errors instead of
silently using a different configurations. If adopters change the
configuration, they want a specific effect and if they mess up, they
should be able to notice that right away and not only several videos are
processed and someone accidentally notices that the setting they made
aren't actually applied.