Remove "chosen" from template dropdown
The "chosen" lib somehow breaks key-value relationship in the template dropdown in the ACL-View. After selecting a template, every subsequently selected template will have the wrong id attached to it. To be precise, each template will have the id from the next template in the dropdown, causing the last template to throw an exception.

What did not work:
- Removing parameters other than "chosen" from the select
- Calling an empty function in ng-change
- Updating "chosen" and "angular-chosen" to newest version