Fallback for Event Title in Publication
The engage publication workflow operation will check in the media
package if a title is set, failing if it is not. This is true, even if
the event has a title set in its main metadata catalog.

The reason for this is that metadata are duplicated at some places and
are not kept in sync properly.

This patch mitigates the issue by looking up the title in the Dublin
Core catalog during publication if non is set in the media package, thus
avoiding a critical failure in this operation.