Fix race condition when loading workflow definitions
There is a race between the OrganizationDirectoryService and
WorkflowDefinitionScanner: workflow definitions that set the organization
field must be loaded after the organization is loaded. If the org is not
loaded beforehand the workflow definition would error out and not be
available. This registers an organization listener. If a new
organization is loaded previously error out workflow definitions are
tried to be loaded in again.