Use bulk inserts for all services during index rebuild
Instead of rebuilding the elastic search index entries with single index requests, now a configurable number of index request are summarized to a bulk request.
Therefor the repopulate functions of all services group multiple documents to a list now and new bundle update functions have been added to the corresponding services and to the elastic search service.