Add optional key "skip-if-target-flavor-has-elements" to Image WOH
Adds an optional key to the image workflow operation handler that skips the operation if the specified target flavor contains at least one element. This can for example be useful if, for some events, you provide the images in some other form and don't want them to be overwritten by your image operation.