Revert try-catch in 12.x MariaDB upgrade script
Among other changes, pull request #4032 added a try-catch clause to the
migration script for migrating MariaDB-based Opencast 11 installations
to Opencast 12. This was later modified in pull request #4101 since it
completely broke the script.

Still, the remaining try-catch means that the script now works
differently than the PostgreSQL variant. It also means that for every
error you get the error message plus stack trace, except for one.

I think not hiding the stack trace is preferable to provide more
information and to assume that all caught `ProgrammingError` have the
same cause may not turn out to be correct.

Therefore, this pull request proposes to remove the try-catch entirely.

This reverts commit 9b78c3362cea01c61ec94bbd3fd22e4256e0eeea.
This reverts commit 9dd7823920813cf262571556215ebf9cf0daead9.