Remove code which is not loading any roles
This patch removes some code which tries to load additional roles for
system users from a file which is hard-coded, does not contain any roles
and cannot be modified by users since it is compiled into the bundle.

In other words, this removes code doing absolutely nothing.

As a nice side-effect, this also removes the Entwine functional library
from this service ;-)