Delete scheduled events fixed (fixes #4084)
Starting with Opencast 12.0 the scheduled events can't be deleted from the Admin UI. As the event can live in more then one place in Opencast, the delete event code try catch them all. Retracting video first, then delte it from scheduler, asset manager, workflow service, index and comment service. But the order is incorrect. The workflow service reads the event ACL. But it was deleted by the scheduler service before. This is the reason of false positive access denied. This patch fixes the event deletion order.