index rebuild indicator and changed rebuild order
The rebuild order has been changed, so that opencast can be used while the workflows are still updated. Since not all information in the admin UI is displayed correctly until the workflow rebuild has finished an index rebuild indicator has been added to the admin UI frontend. This indicator is getting the states of the repopulation service from the IndexRebuildService using a new endpoint and shows them to the user within the header of the admin UI. If no rebuild is running the indicator is hidden.