Make busy waiting explicit during InboxScannerService config update
I have now been bitten twice by this so let's fix this: the
`InboxScannerService` could busy wait forever if the configured
user is not available (e.g. you don't configure the default admin user).
This can hinder other services from properly configuring themselves.
This change introduces a limit and makes it very explicit in the log
that it is waiting.

The proper solution for adopters not using the inbox is to remove the
configuration file. This code will then not be executed. Of course I
forgot that when updating to Opencast 12.