Added automatic language detection to Azure Transcription Service
The Speech-to-Text Service by Microsoft Azure offers to detect the spoken language for you, allowing you to request transcription without setting a specific language. This PR allows users to use that functionality through global configuration or WOH keys.

The detected language is also stored in a file in the workspace, so the Attach WOH can later know which language was detected and set the file flavor to include the language code if the user so wishes. It does not set it as a workflow property because the transcription (and thus the language detection) run asynchronously, so technically it is uncertain whether the workflow which started the transcription is still active when the language detection is done. Still, this is quite a roundabout way of getting the language to the Attach WOh, so if there are better ways to solve this problem (or if my explanation is just wrong), please let me know.