Update location field in archive when CA is changed
The scheduler service updates this field in the index when the capture agent is
changed, but doesn't actually make a snapshot, so this change is
overwritten when the next snapshot is made.
This is also relevant because this field is used in the live schedule
service to generate the URLs for the live tracks.