Join index rebuild of scheduler to assetmanager
The repopulate function from the scheduler service has been removed. The index rebuild of scheduled events will now be proceessed as a part of the asset manager rebuild. Therefor a new SchedulerIndexData Object will be created, selecting the relevant data from the scheduler database table. Within the rebuild of the assetmanager the scheduler data is checked for the given mediapackage id. If the event is a scheduled event it will by extended by the scheduler data, if not it will be processed as before.