Add captions to Tobira harvest API
Adds a field `captions` to the `event` item. All attachments and tracks
are considered a caption if they have the flavor `caption/vtt*` or the
MIME type `text/vtt`. If the flavor has the form `caption/vtt+<lang>`
where `<lang>` could be anything, then `<lang>` is included in the
harvest response (null otherwise).