Make Workflow Operation Docs Look Alike
This patch starts adjusting the workflow operation handler documentation
to always use a similar template, starting with the operations name and
it's identifier and, not using the Java class name and ending with an
example in a similar format with syntax highlighting enabled.

This does not yet update all workflow operation handler docs.