WebVTT subtitle editing in the video editor backend (#3606)
This patch modifies the video editor service to be able to cut subtitle tracks based on the WebVTT format.

As FFmpeg is severely limited in how it allows cutting subtitles (does not work with filters), this parses the WebVTT subtitles into a list container to allow for fine-grained editing by the video editor. The WebVTT parser is still rather rudimentary, but should not fail as long as the given WebVTT file is valid.

The video editor now discriminates between the usual audio/video elements and reference elements (aka everything that is not explicitly specified to be an audio or video element). Within the reference elements, it looks for `.vtt` files and cuts them according to the SMIL.