Specify Units in Video Clips
The `VideoClip` class stores a begin and an end value but does not
specify the unit these values are stored in. This lead to code storing
the values in seconds while in the next line code would store
milliseconds. That is really confusing.

This patch now makes it clear that all input values are in milliseconds
since that is what our code uses anyway and we have specific functions
to get these vales either as seconds or as milliseconds.

This hopefully makes the code less confusing to read.