Fix saving thumbnails if no previous ones existed
This patch fixes a bug in the editor backend causing the editor to fail
saving data if users set a new thumbnail and no previous thumbnails
existed.

The problem was that the code tried to get the tags from previous
thumbnails and assumed that at least one would always exist.