Automate Release Process
Right now, cutting a new release is mostly just copy and pasting
commands from the developer documentation, hoping that you aren't so
bored that you miss changing a version number in those commands.

This patch completely automates the release process. To cut a new
release, a release manager can now go to GitHub Actions and trigger the
“Cut Release Tag” workflow.

The GitHub Actions user interface will ask for a branch to run this on
and the release manager would select the release branch, e.g. `r/13.x`.
The workflow will check this and automatically fail on non-release
branches.

Based on the release branch, the workflow will then extract the major
version and based on existing tags automatically determine the next
version in line.

For example, selecting `r/13.x`, with no 13.x versions tagged right now,
the workflow will determine that the 13.0 release needs to be cut.

The workflow will then update the `pom.xml` versions, build Opencast and
check it starts up properly, commit the changes, create a git tag and
finally push the tag to GitHub. In other words, it does all the annoying
stuff a release manager would usually have to do manually.