Correctly identify the master playlist in the Tobira harvest
Some events in Opencast (for example those processed by
the `multiencode` operation) have multiple HLS streams in their
media package. These represent different levels of quality as far
as I understand. In this case there will be one among them makred
as "master," which refers to the others.

Paella (7 at least) only ever expects and works with one (the first)
HLS stream, so for quality switching to work in the above situation
we have to make sure to pass said "master track." However, Tobira
currently has no clue which one that is.

This patch exposes the required information in Tobira's
harvesting API, allowing us to make the required distinction in
a follow up PR on the Tobira side.

This change is not breaking in any way: Existing Tobiras will just
ignore the information. Without it, they will however
exhibit elan-ev/tobira#573.