Allow use of extended metadata in send-email WOH
Currently, the send-email WOH simply searches for catalogs with the
flavor `dublincore/*` to populate the flavor. This circumvents the
catalog configuration and does not allow for flavors with a different
main type.

This now uses the index service that already has logic for filtering for
configured catalogs of the current tenant. Only configured catalogs are
considered, i.e. this could be a breaking change if adopters use e.g.
`dublincore/foo` in the mediapackage but never configure it as a
catalog. I think this is a configuration error.

To allow for multiple catalogs with the same subflavor, the templating
now uses the full flavor name, e.g. `${catalogs['dublincore/episode']['title']}`.
As a special case for `dublincore/*`, it is still posible to only use
the subtype. I would declare this as deprecated.